App is now written in go and can utilise all cores

+++ bin/app.sh
@@ -5,7 +5,11 @@ set -e
 APP_NAME=$1
 
 function delete_cgroup {
- while true; do
+  if [ ! -d "/sys/fs/cgroup/cpu/$1/$APP_NAME" ]; then
+    return
+  fi
+
+  while true; do
    if rmdir "/sys/fs/cgroup/cpu/$1/$APP_NAME" 2> /dev/null; then
      break
    fi
@@ -18,7 +22,22 @@ function cleanup {
  echo $$ > /sys/fs/cgroup/cpu/tasks
  delete_cgroup "good"
  delete_cgroup "bad"
+ exit
 }
+
+function eat_cpu {
+  while true; do
+    if [ ! -f "$1/spike" ]; then
+      sleep 5
+    else
+      for i in $(seq 1000000); do
+        true
+      done
+    fi
+  done
+}
+
+export -f eat_cpu
 
 trap cleanup EXIT
 
@@ -31,12 +50,4 @@ mkdir -p "/sys/fs/cgroup/cpu/good/$1"
 echo $$ > "$APP_DIR/pidfile"
 echo $$ > "/sys/fs/cgroup/cpu/good/$1/tasks"
 
-while true; do
-  if [ ! -f "$APP_DIR/spike" ]; then
-    sleep 1
-  else
-    for i in $(seq 1000000); do
-      true
-    done
-  fi
-done
+parallel -n0 eat_cpu ::: $APP_NAME $APP_NAME
